Added buggy initial try at build system. Still needs to be written for windows/linux, documented and tested. (It does not really work properly at the moment, even on OS X) Made minor fixes to code elsewhere. Made useless changes to main.cpp

+++ PauseInitiatorID.h
@@ -20,7 +20,7 @@ namespace hg {
     {
     public:
         PauseInitiatorID(pauseinitiatortype::PauseInitiatorType type,
-                         unsigned int ID,
+                         size_t ID,
                          size_t timelineLength);
         PauseInitiatorID() :
         type_(pauseinitiatortype::INVALID),
@@ -28,7 +28,7 @@ namespace hg {
         timelineLength_(0)
         {}
         pauseinitiatortype::PauseInitiatorType type_;
-        unsigned int ID_;
+        size_t ID_;
         size_t timelineLength_;
         friend class boost::serialization::access;
